Renames mixin directory to mixins.
