+++ myuw/static/vue/mixins/outlink.js
